🚚 [yw_clients] Gather `aiohttp` features in single file
